Fix Linker errors in Titanium
