+++ JavaScriptCore/runtime/MapIteratorConstructor.cpp
@@ -38,7 +38,7 @@
 #include "JSGlobalObject.h"
 #include "JSMapIterator.h"
 #include "MapIteratorPrototype.h"
-
+#include "StructureInlines.h"
 namespace TI {
 
 const ClassInfo MapIteratorConstructor::s_info = { "MapIterator Iterator", &Base::s_info, 0, 0, CREATE_METHOD_TABLE(MapIteratorConstructor) };
